requirements.txt
pyjwt==1.7.1
django==3.2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
django-filter==21.1